Fix remarks.Refactor operation getEntity for one transaction action.

+++ repository-module/src/main/java/com/gmail/kaminski/viktar/onlinemarket/repository/GenericRepository.java
@@ -14,5 +14,5 @@ public interface GenericRepository<I, T> {
 
     List<T> findAll(int firstElement, int amountElement);
 
-    Long getAmountOfEntities();
+    Integer getAmountOfEntities();
 }
